+++ movie/src/main/java/com/kh/movie/dao/ReviewDao.java
@@ -21,5 +21,6 @@ public interface ReviewDao {
 	List<ReviewListVO> findByRatingAsc(int movieNo);//평점낮은순
 	
 	List<MovieSimpleInfoDto> findAll(int movieNo);//영화 정보 조회
-	int findReviewNoByMovie(int movieNo);//영화에 달린 리뷰 조회
+	List<ReviewListVO> findReviewNoByMovie(int movieNo);//영화에 달린 리뷰 조회
+	int findReviewLikeCount(int reviewNo);//좋아요 개수 조회
 }
