--- movie/src/main/java/com/kh/movie/vo/ReviewLikeResultVO.java
@@ -1,12 +1,0 @@
-package com.kh.movie.vo;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data @NoArgsConstructor @AllArgsConstructor @Builder
-public class ReviewLikeResultVO {
-	private int reviewNo;
-	private String reviewLike;
-}
