--- movie/src/main/resources/mybatis/review-list-mapper.xml
@@ -1,50 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!DOCTYPE mapper
-    PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-    "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-
-<mapper namespace="reviewList">
-
-	<!-- 최신순 -->
-	<select id="findByDateDesc" resultType="ReviewListVO">
-		select * from review_list where movie_no = #{movieNo} order by review_date desc
-	</select>
-	
-	<!-- 오래된순 -->
-	<select id="findByDateAsc" resultType="ReviewListVO">
-		select * from review_list where movie_no = #{movieNo}  order by review_date asc
-	</select>
-	
-	<!-- 좋아요순 -->
-	<select id="findByLikeDesc" resultType="ReviewListVO">
-		select * from review_list where movie_no = #{movieNo}  order by review_like_count desc
-	</select>
-	
-	<!-- 평점높은순 -->
-	<select id="findByRatingDesc" resultType="ReviewListVO">
-		select * from review_list where movie_no = #{movieNo}  order by rating_score desc
-	</select>
-	
-	<!-- 평점낮은순 -->
-	<select id="findByRatingAsc" resultType="ReviewListVO">
-		select * from review_list where movie_no = #{movieNo}  order by rating_score asc
-	</select>
-	
-	
-	<!-- 상세조회 -->
-	<select id="findByReviewNo" resultType="ReviewListVO">
-		select * from review_list where review_no = #{reviewNo}
-	</select>
-
-	<!-- 영화에 달린 리뷰 조회 -->
-	<select id="findReviewNoByMovie" resultType="int">
-	    select review_no from review_list where movie_no = #{movieNo}
-	</select>
-	
-	<!-- 좋아요 개수 조회 -->
-	<select id="findReviewLikeCount" resultType="int">
-		select count(*) from review_like where review_no = #{reviewNo}
-	</select>
-
-</mapper>
